fix: update message/media controller error message (#25)
* fix(media_exchange): update media_exchange description

* feat(media_exchange): add scenario "compartilhamento de midia"

* feat(media_exchange): add scenario "falha no compartilhamento de midia"

* fix: updates media removal scenario

* fix: updates media receival scenario

* fix(media_exchange): specify error condition for media sending

* fix: remove mixed gui and service scenario

* feat: add failed media removal scenario

* fix(media_exchange): alter redundant language and split removal scenario

* fix: make npm start script able to run

* fix: re-add .env.example file

* feat: add socket.io server

* feat: add html code for better visualization of backend requests

* feat: allow app to use public as its static path

* feat: add submit button and message container to static html page

* feat: add message server and client

* feat: allow multiple clients to connect

* feat: add basic message sending

* feat: make socket connection accept file transfer

* feat: display transfered file

* feat: render different types of media (pictures, audio, video and documents)

* feat: add chat interface, schema and model

* fix: delete unused files

* refactor: change contacts.interface location

* fix: remove unused import

* feat: add file writing to server

* feat: make message/media transfer use http instead of sockets

* feat: make message/media transfer use http instead of sockets

* feat: add messages database

* fix: make submit button send sender's name

* feat: add message/media deletion request

* fix: re-add .env.example

* feat: add service scenarios and getMediaConversation method

* fix: update chatSchema

* feat: update service scenarios

* feat: add media sending tests

* fix: change error message for media get

+++ backend/src/controllers/message.controller.ts
@@ -73,16 +73,18 @@ class MessageController {
     const messageId = req.params.messageId;
     const messageSent = this.database.getMessage(messageId)
 
+    if (!messageSent?.id) {
+      return new FailureResult({
+          msg: 'Message not found',
+      }).handle(res);
+    }
+    
     if (messageSent?.sender === req.params.sender) {
       const success = this.database.deleteMessage(messageId);
       if (success) {
         return new SuccessResult({
             msg: 'Message deleted successfully',
         }).handle(res);
-      } else {
-          return new FailureResult({
-              msg: 'Message not found',
-          }).handle(res);
       }
     } else {
       return new FailureResult({
